chore: #227 1차 코드리뷰 반영
